requirements.txt
asgiref==3.3.1
Django==3.1.7
django-appconf==1.0.4
django-tinymce==3.2.0
django-tinymce-lite==0.0.2
django-tinymce4-lite==1.8.0
jsmin==2.2.2
phonenumbers==8.12.20
Pillow==8.1.2
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
